move 'snippets' related code to own module
